tls: move legacy code into own file
PR-URL: https://github.com/nodejs/node/pull/39333
