+++ lib/_tls_common.js
@@ -52,8 +52,11 @@ const {
 
 const {
   configSecureContext,
+} = require('internal/tls/secure-context');
+
+const {
   parseCertString,
-} = require('internal/tls');
+} = require('internal/tls/parse-cert-string');
 
 function toV(which, v, def) {
   if (v == null) v = def;
